// online-quiz-maker-main/client/src/context/authContextValue.ts
import { createContext } from "react";

interface User {
  id?: string;
  name: string;
  email: string;
  tests?: unknown[];
}

export interface AuthContextType {
  user: User | null;
  token: string | null;
  login: (token: string, user: User) => void;
  logout: () => void;
}

export const AuthContext = createContext<AuthContextType>({
  user: null,
  token: null,
  login: () => {},
  logout: () => {},
});